Moving test files, utilizing utility for noop
