+++ packages/story-editor/src/components/floatingMenu/elements/karma/shopping.karma.js
@@ -67,70 +67,6 @@ describe('Shopping integration', () => {
     return storyContext.state.selectedElements[0];
   };
 
-  describe('Shopping tab', () => {
-    it('should handle product search add and remove', async () => {
-      const productTitle = 'Hoodie';
-      await focusProductSearchInput();
-      await fixture.events.keyboard.type('hood');
-      // delay for search to catch-up
-      await fixture.events.sleep(400);
-
-      const productButton = fixture.querySelector(
-        `[aria-label="Add ${productTitle}"]`
-      );
-      await fixture.events.click(productButton);
-      await waitFor(() =>
-        fixture.querySelector(
-          '[aria-label="Design menu"] [aria-label="Product"]'
-        )
-      );
-
-      // add a small delay for debounce search to catchup
-      await fixture.events.sleep(500);
-
-      // check story `state`
-      const selectedElement = await getSelectedElement();
-      await expect(selectedElement?.product?.productTitle).toBe(productTitle);
-      await focusProductSearchInput();
-
-      // select the product add / remove button
-      const product = fixture.querySelector(
-        `[aria-label="Remove ${productTitle}"]`
-      );
-
-      // check add / remove icons
-      const checkIcon = fixture.querySelector(
-        `[aria-label="Remove ${productTitle}"] svg[class^="productButton__Checkmark-"]`
-      );
-
-      const removeIcon = fixture.querySelector(
-        `[aria-label="Remove ${productTitle}"] svg[class^="productButton__Cross-"]`
-      );
-
-      expect(window.getComputedStyle(checkIcon).display).toBe('block');
-      expect(window.getComputedStyle(removeIcon).display).toBe('none');
-      await fixture.events.hover(product);
-      expect(window.getComputedStyle(checkIcon).display).toBe('none');
-      expect(window.getComputedStyle(removeIcon).display).toBe('block');
-
-      // remove the product
-      await fixture.events.mouse.clickOn(product, 1, 1);
-    });
-
-    it('should disable button if product lacks product image', async () => {
-      await focusProductSearchInput();
-      await fixture.events.keyboard.type('WordPress');
-      // delay for search to catch-up
-      await fixture.events.sleep(400);
-
-      const productButton = fixture.querySelector(
-        '[aria-label="Products without images cannot be added."]'
-      );
-
-      await expect(productButton.getAttribute('aria-disabled')).toBe('true');
-    });
-  });
-
   describe('Product floating menu', () => {
     it('should render products menu', async () => {
       const productTitle = 'Album';
